Create README.md
READ ME
To talk to Leo, you must use 'character name: your message'
Sophia is set as the default character for you to play.
You are able to change the default character in the prompt, however, this is not advised.

To have a better experience talking to Leo, use '(what the character you are playing does or did)' for example, (giggles)/(stands up).

Conversations can go for up to 7 messages from either end. 

You may or may not regenerate responses if Leo breaks before the 7 message point.

If for any reason you would like to keep the conversation going, you must copy your whole conversation and put it as the new 'scenario'.

To have the best experience with Leo, you must start every sentence under 'scenario' with 'Leo' and try to mention 
Also, mention Leo in your messages as much as possible to make sure Leo replies and not chatgpt.

Enjoy!